Time: 3 ms (61.07%) | Memory: 6.6 MB (71.77%) - LeetSync

+++ 8-string-to-integer-atoi/string-to-integer-atoi.cpp
@@ -29,7 +29,7 @@ public:
         }
 
         // Convert string to integer
-        long long s_long;
+        long s_long;
         std::stringstream(s_clean) >> s_long;
 
         // Handle negative sign
